Move Sprite class to SpriteSheet class.
